requirement.txt
langchain
python-dotenv
langchain_core
langchain_community
streamlit